move showModal helper into its own pack file
